smart_farm_backend/requirements.txt
fastapi
uvicorn
sqlalchemy
paho-mqtt
python-dotenv
